revert NotebookRepoCommon
