+++ zeppelin-zengine/src/main/java/org/apache/zeppelin/notebook/Note.java
@@ -47,7 +47,7 @@ import org.apache.zeppelin.interpreter.InterpreterSetting;
 import org.apache.zeppelin.interpreter.InterpreterSettingManager;
 import org.apache.zeppelin.interpreter.remote.RemoteAngularObjectRegistry;
 import org.apache.zeppelin.interpreter.thrift.InterpreterCompletion;
-import org.apache.zeppelin.notebook.repo.NotebookGitRepo;
+import org.apache.zeppelin.notebook.repo.NotebookRepoWithVersionControl;
 import org.apache.zeppelin.notebook.repo.NotebookRepo;
 import org.apache.zeppelin.notebook.utility.IdHashes;
 import org.apache.zeppelin.scheduler.Job;
@@ -303,7 +303,7 @@ public class Note implements ParagraphJobListener, JsonSerializable {
   }
 
   void setRevisionSupported(NotebookRepo repo) {
-    if (repo instanceof NotebookGitRepo) {
+    if (repo instanceof NotebookRepoWithVersionControl) {
       getConfig().put("isRevisionSupported", true);
     } else {
       getConfig().put("isRevisionSupported", false);
